added some javascript to make form tabs active

+++ lib/delete.php
@@ -2,6 +2,7 @@
 if (
   isset($_GET['action']) && $_GET['action'] == 'delete'
   && isset($_GET['fname'])
+  && isset($_GET['path'])
 ) {
   $filePath = $_GET['path'] . '/' . $_GET['fname'];
   unlink("./{$filePath}");
